Switch from int32_t to size_t for variables which get assigned a value returned from strlen() and which are compared to such values.

+++ sketch/Billy/cstring_utils.cpp
@@ -30,7 +30,7 @@ int32_t cstring_utils::to_lowercase(char *str)
     }
 
     int32_t i = 0;
-    for (int32_t j = 0, len = strlen(str); j < len; ++j) {
+    for (size_t j = 0, len = strlen(str); j < len; ++j) {
         if (str[j] >= 'A' && str[j] <= 'Z') {
             str[j] += 'a' - 'A';  // Difference between an ASCII code of an uppercase and a lowercase letter.
             ++i;
@@ -47,7 +47,7 @@ int32_t cstring_utils::to_uppercase(char *str)
     }
 
     int32_t i = 0;
-    for (int32_t j = 0, len = strlen(str); j < len; ++j) {
+    for (size_t j = 0, len = strlen(str); j < len; ++j) {
         if (str[j] >= 'a' && str[j] <= 'z') {
             str[j] -= 'a' - 'A';  // Difference between an ASCII code of an uppercase and a lowercase letter.
             ++i;
@@ -63,7 +63,7 @@ int32_t cstring_utils::nullify_first_cr_or_lf(char *str)
         return CSTRING_UTILS_MEM_ERR;
     }
     
-    for (int32_t i = 0, len = strlen(str); i < len; ++i) {
+    for (size_t i = 0, len = strlen(str); i < len; ++i) {
         if (str[i] == '\r' || str[i] == '\n') {
             str[i] = '\0';
 
@@ -80,14 +80,16 @@ int32_t cstring_utils::nullify_trailing_crs_and_lfs(char *str)
         return CSTRING_UTILS_MEM_ERR;
     }
 
-    int32_t len = strlen(str);
+    size_t len = strlen(str);
     if (len == 0) {
         return CSTRING_UTILS_NOT_PROCESSED;
     }
 
     int32_t i = 0;
-    while (len > 0 &&  // len can decrement during the loop execution, therefore another check is done.
-           (str[len - 1] == '\r' || str[len - 1] == '\n')) {
+    /* len can decrement during the loop execution,
+     * threfore it is checked to prevent an underflow.
+     */ 
+    while (len > 0 && (str[len - 1] == '\r' || str[len - 1] == '\n')) {
         str[len - 1] = '\0';
         --len;
         ++i;
@@ -103,7 +105,7 @@ int32_t cstring_utils::nullify_all_crs_and_lfs(char *str)
     }
 
     int32_t i = 0;
-    for (int32_t j = 0, len = strlen(str); j < len; ++j) {
+    for (size_t j = 0, len = strlen(str); j < len; ++j) {
         if (str[j] == '\r' || str[j] == '\n') {
             str[j] = '\0';
             ++i;
@@ -119,7 +121,7 @@ int32_t cstring_utils::append_cr(char *str, size_t arr_size)
         return CSTRING_UTILS_MEM_ERR;
     }
 
-    int32_t len = strlen(str);
+    size_t len = strlen(str);
     if (arr_size - len < 2) {  // One byte for an appended character, another byte for null.
         return CSTRING_UTILS_MEM_ERR;
     }
@@ -136,7 +138,7 @@ int32_t cstring_utils::append_lf(char *str, size_t arr_size)
         return CSTRING_UTILS_MEM_ERR;
     }
 
-    int32_t len = strlen(str);
+    size_t len = strlen(str);
     if (arr_size - len < 2) {  // One byte for an appended character, another byte for null.
         return CSTRING_UTILS_MEM_ERR;
     }
@@ -153,7 +155,7 @@ int32_t cstring_utils::append_char(char *str, size_t arr_size, char char_to_appe
         return CSTRING_UTILS_MEM_ERR;
     }
 
-    int32_t len = strlen(str);
+    size_t len = strlen(str);
     if (arr_size - len < 2) {  // One byte for an appended character, another byte for null.
         return CSTRING_UTILS_MEM_ERR;
     }
@@ -170,14 +172,16 @@ int32_t cstring_utils::count_trailing_crs_and_lfs(char *str)
         return CSTRING_UTILS_MEM_ERR;
     }
 
-    int32_t len = strlen(str);
+    size_t len = strlen(str);
     if (len == 0) {
         return 0;
     }
 
     int32_t i = 0;
-    while (len > 0 &&  // len can decrement during the loop execution, therefore another check is done.
-           (str[len - 1] == '\r' || str[len - 1] == '\n')) {
+    /* len can decrement during the loop execution,
+     * threfore it is checked to prevent an underflow.
+     */ 
+    while (len > 0 && (str[len - 1] == '\r' || str[len - 1] == '\n')) {
         --len;
         ++i;
     }
@@ -195,7 +199,7 @@ int32_t cstring_utils::append_lf_if_no_trailing_crs_nor_lfs(char *str, size_t ar
         return CSTRING_UTILS_NOT_PROCESSED;
     }
 
-    int32_t len = strlen(str);
+    size_t len = strlen(str);
     if (arr_size - len < 2) {  // One byte for an appended character, another byte for null.
         return CSTRING_UTILS_MEM_ERR;
     }
